LPS-95418 transit generator to spi
