+++ modules/apps/portal-security-credential/portal-security-credential-generator-bcrypt/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Security Credential Generator BCrypt
+Bundle-SymbolicName: com.liferay.portal.security.credential.generator.bcrypt
+Bundle-Version: 1.0.0
+Liferay-Require-SchemaVersion: 3.0.0
