+++ modules/apps/portal-security-credential/portal-security-credential-generator-spi/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Portal Security Credential Generator SPI
+Bundle-SymbolicName: com.liferay.portal.security.credential.generator.spi
+Bundle-Version: 1.0.0
+Liferay-Require-SchemaVersion: 3.0.0
