+++ modules/apps/portal-security-credential/portal-security-credential-generator-spi/build.gradle
