--- modules/apps/portal-security-credential/portal-security-credential-generator/src/main/java/com/liferay/portal/security/credential/generator/BCryptCredentialGenerator.java
@@ -1,49 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.security.credential.generator;
-
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.security.credential.model.AlgorithmEntry;
-import com.liferay.portal.security.credential.model.CredentialMeta;
-import com.liferay.portal.security.crypto.generator.salt.BCryptSaltGenerator;
-import com.liferay.portal.security.crypto.generator.secret.BCryptSecretGenerator;
-
-/**
- * @author arthurchan35
- */
-public class BCryptCredentialGenerator {
-
-	public static String generateSalt(AlgorithmEntry algorithm)
-		throws Exception {
-
-		JSONObject meta = JSONFactoryUtil.createJSONObject(algorithm.getMeta());
-
-		int rounds = meta.getInt("rounds");
-
-		if (rounds < 1) {
-			throw new Exception();
-		}
-
-		return BCryptSaltGenerator.generate(rounds);
-	}
-
-	public static String generateSecret(String password, CredentialMeta meta)
-		throws Exception {
-
-		return BCryptSecretGenerator.generate(password, meta.getSalt());
-	}
-
-}
